category-games: add multiple domains
Modrinth is a mod distribution platform.

Nexus Mods is a platform for people around the world to share
their mods and tools freely with other gamers.

PaperMC is a Minecraft Software organization focusing on improving the
Minecraft ecosystem with faster and more secure software.

ProtonDB gathers reports from other gamers as they test games with
Proton on Linux and provide aggregate scores of how well games perform.

speedrun.com is a site that provides leaderboards, resources, forums and
more, for speedrunning.

SteamDB is a third-party website gives you better insight into the Steam
platform and everything in its database.