Add multiple domains
Zola is a fast static site generator in a single binary with everything
built-in.

ShellCheck is a static analysis tool for shell scripts.

Bromite is a Chromium fork with ad blocking and privacy enhancements.

Touhou Wiki is dedicated to the Touhou Project, a Japanese gaming series
that mostly contains curtain fire shoot'em up games made by ZUN.