Update test-pr.yml
I think we can provide the file which has been generated to the contributor. It has two reasons. First, the contributor can get the file as soon as possible. At the same time, it will not influence other users before we merge it. Second, the contributor will test the part which has been changed by himself/herself. If he/she gets any problem, he/she can fix it before we find it.